ENH: Include optimizer attribute of type IOptimizer on APosterioriOptimizer and WeightedOptimizerEngine classes
